server: Fade in looped sounds
(Code from both Philip & Fabián)

This creates a control binding to the pipeline's volume element, that we
can add control points to, to implement the fade-in.

The volume element is now added unconditionally to the pipeline. This
change is not strictly necessary, but if we want to support in the
future the ability to change the volume of sounds in-flight, then we
have to do this anyway, and it makes the code simpler.

The fade-in control points are added when a looped sound is started, and
removed when the sound loops back to the beginning.

https://phabricator.endlessm.com/T24494