server: Support updating properties on the fly among loops
Before this commit, keyframes applied to a higher position than
the end of the stream didn't have effect. If an audio file had 5
seconds, and if there was a request to add a keyframe between
time=4s and time=6s, firstly the update of the properties was
considered only between time=4s and time=5s and secondly the
update of the properties would be looped between time=4s
and time=5s.

The problem with the last approach was that sometimes it may be
the case that updating the properties of the pipeline may not
"sound" good and you hear brusque variations in the sound. For
example, it may have been requested to update the volume property
from rate=1.0 to rate=2.0 from time=4s to time=6s. In the last
approach, we would have stoped updating the rate property at
time=5s (considering a sound file duration of 5 seconds), and thus
the rate would be equal to 1.5. In the next loop, the rate would
start again from rate=1.0. There would be an abrupt change of the
rate from rate=1.5 to 1.0.

The problem described above was caused basically because the clock
was restarted at the end of each loop. In the case of the sound
with a duration of 5 seconds, the position of the pipeline after
time=5s would be time=0s. In this approach, after time=5s, the
clock keeps increasing so in the next seconds the position would
be time=6s, time=7s, time=8s... and so on. This is also useful
because we don't have the need to split keyframes.

This commit avoids that the EOS event gets emmited at the end of
each loop and instead uses SEGMENT seeks so the streaming clock
gets restarted to zero at the end of each loop. This task should be
just  achieved by the use of SEGMENT seeks, however it seems that
the decodebin is not handling properly the segment seeks, and the
clock isn't restarted Adding the 'identity' helps to fix the
buffers timestamps so the clock doesn't get restarted.