metadata: Use multiple sounds
Now, an entry in the metadata can be something like this.

    "sound-files": [
        "sound1.wav",
        "sound2.wav"
    ]

If a list of sounds in "sound-files" the server will pick one
randomly. If only a sound "sound-file" is given it will be
converted to a "sound-files" list of an only one item. If both
"sound-file" and "sound-files" are given, their values will be
merged into "sound-files".