server: Change the behavior of PlaySound
A new property "overlap-behavior" is added to the metadata file
so the playback of an audio could be restarted or ignored.

    "example": {
        "sound-file": "example.wav",
        "overlap-behavior": "overlap"
    }

The default behavior is "overlap" even if "overlap-behavior" is
not specified. In this case, a new UUID is created every time
PlaySound is called. If "reset" is specified, then if the sound
is playing, it will be seek to it start point restarting all
keyframe values. If "ignore" is specified, then if the sound is
playing, and another call to PlaySound comes in, this call will
just be ignored. In both last cases, the existent UUID will be
returned.