server: Ignore "ignore" sound just if its uuid is registered
In the last implementation, it was checked that the state of the
sound should be equal to the PLAYING state. However, the state
may going through NULL -> READY -> PAUSED which would cause to
not ignore the sound. We just assume that a registered uuid will
be always matched to a playing pipeline, so if the uuid is
registered that is enough to know if the sound should be ignored
or not.