metadata: Error in metadata should not be fatal
In the case where the user metadata is wrong, we can just fall back to
the system metadata. In the case where the system metadata is wrong, we
can't really play any sounds, but the existing behaviour was already to
ignore JSON decode errors but crash on metadata validation errors; this
makes it consistent instead.

https://phabricator.endlessm.com/T29444