server: Stop looping sound as soon volume reaches 0
Previously, looping sounds waited for the last loop to finish. So,
for example, Let's imagine a looping sound durated 10 seconds
with 1 second duration for the fade-out effect. If a stop call
was sent at t=4s, then the volume would reach volume=0 at t=5s,
but the sound was stopped at t=10s. For this same example, the
new added change, would make to stop the sound t=5s.

https://phabricator.endlessm.com/T24852