server: Support priority for loop events
A priority loop is an event that precedence over
other loops. This means that:

a) a non-priority loop will be paused when a priority loop is played.
b) a non-priority loop will be started in a paused state if a
   priority loop is already playing.
c) paused loops will be resumed after the priority loop stops.

This version doesn't support different priority levels, therefore
two priority loops can be active at the same time.

https://phabricator.endlessm.com/T25022