server: Allow to unset values before setting a keyframe pair
Before this change, a call to UpdateProperties with a transition
time, while a previous transition was still in progress, would
leave the original keyframes intact, causing unexpected behaviour.
For example, if fading out the volume while a longer fade-in was
in progress, the volume might jump from 0 to maximum when the end
of the original fade-in was reached.

https://phabricator.endlessm.com/T25046