server: Support pausing or silencing a sound when another plays
Now the metadata for a sound event id can defined something like:

    "sound-x": {
        "apply-state-on": {
            "pause": ["soundA", "soundB"],
            "silence": ["soundC", "soundD"]
        }
    }

This means that whenever "sound-x" is played the states "pause"
would be applied to "soundA" and "soundB" and the state "silence"
would be applied to "soundC" and "soundD". And whenever the
"sound-x" is resumed the paused/silenced sounds should resume with
a fade effect.

Warning:
Sounds are resumed to the default volume but the volume may have
changed in between using the method UpdateProperties.

https://phabricator.endlessm.com/T25155