server: PlayFull: Calculate property values following a function
Instead of just multiplying by a value, this commit allows the
user to evaluate an input "x" value against an f(x) function.
This f(x) function is an estimated polynomial taking as input
a "degree" and "points" defined in the metadata.

Use as following:

    ```
    "my/sound": {
        "sound-file": "whatever.wav",
        "recalculate": {
            "pitch": {
                "degree": 1,
                "points": [
                    [0, 2],
                    [1000, 0.5]
                ],
                "min": 0.5,
                "max": 2
		"x": 5000
            }
        }
    }
    ```

If PlayFull is called with options
    {"recalculate": {"pitch": {"x": 800}}}}
then the function will evaluate f(800).