server: Limit to 5 sounds playing simultaneously
It avoids causing bottlenecks on the sound server when too many
requests come in a very short time.

The rule is the following: if the user attempts to create a new
instance/uuid of sound event id that already have 5 instances,
then the request will be ignored and considered as an error.

https://phabricator.endlessm.com/T25273