server: Silence sounds when its window is unfocused
Two properties have been added to the metadata:
    - "target-app-id": The dbus bus well-known-name of the app
                       using the sound
    - "side": Whether the sound is pretended to be played in the
              "frontside" or the "backside"

If no "target-app-id" is specified then the rule to silence sounds
when "its window is unfocused" will be totally ignored.

If in the backside, frontside sounds will be silence and backside
sounds will play. The other way around if in the frontside.

https://phabricator.endlessm.com/T25022