server: Avoid pausing a sound if being released or stopping
References to a player is registered in multiple dictionaries
(for example, the HackSoundServer.players dictionary). While a
player is registered in these dictionaries, users/clients can
call methods that apply actions on these players. Players are
not always released from these dictionaries as soon the StopSound
method is called but when the fade-out effect (if any) finishes.

This commit, prevents that an action like pausing the player
happened while a sound is being released or stopped. Not doing
that would be bad for three indentified reasons:

- Because the sound or player gets paused it could never complete
  the fade out effect. So the release method wouldn't be called.
- Not releasing the players would imply wasting of memory.
- In the case of overlapping sounds, because sounds are never
  released, it is probable that the server ends up with 5 paused
  sounds causing that all the following requests to play the same
  sound event id get rejected because they are still referenced and
  the server has a limit of 5 simultaneous sounds.

  One may argue that this behavior already occurs when sounds are
  fading out, but there is no decision about that yet, and this is
  not too bad, because after finishing the fade effect it will be
  possible to play back sounds of the same sound event id.
- In the case of non-overlapping sounds, because sounds are set to
  PLAYING state only when the reference count is 1, they will be
  never have the chance to play again.

https://phabricator.endlessm.com/T25659