server: Avoid playing if releasing
Previously, because the method to release is the player is invoked
through GLib.idle_add something may have be called just before it
the HackSoundServer._release method gets executed. For example,
the PlaySound method could be invoked "just before" the player
gets released.

This change prevents to play a sound if the player is getting
released.

https://phabricator.endlessm.com/T25659