server: Avoid server crash due to unhandled exception on pausing
Previously, if a call to StopSound arrived very quick just after
a PlaySound call, the server would have crashed because an
unhandled exception.

This problem was because a StopSound call may have arrived while
the player is still in NULL or READY applying a fade out effect
which depends on the current pipeline position, but according the
documentation a GST_QUERY_POSITION query should usually work only
if the pipeline is in PLAYING or PAUSED state.

This commit avoids the crash by just handling the exception.

https://phabricator.endlessm.com/T25659