server: Play with fade in (if specified) always
Previously, PlaySound calls were only faded in when the refcount
was equal to 1 for non overlapping sounds. This condition was
added because in the Clubhouse a new quest notification dialog
may have requested to play the same (uuid of the) previous sound,
causing the sound to fade in again (which was not undesired). It
produced a bad effect, because before, fade in effects were always
applied from volume 0.

This change removes that previous condition to allow sounds to
always fade in when a call to PlaySound. The bad effect wouldn't
occur because the fade in effect has been changed in the previous
commit to fade in from the current volume. So in the case of the
Clubhouse quests it would fade its volume from 1 to 1.

Note:
    Somebody may argue that if the current volume is greater than
    the default volume, the sound would fade out instead of fading
    in. But there is no decision on that yet.

https://phabricator.endlessm.com/T25659