server: Allow to play a sound even if it is stopping
Previously, sounds weren't allowed to be played back if a fade out
effect was in progress for non-overlapping sounds. It means that
PlaySound calls on a stopping sound didn't have effect until the
player got released (when the fade out effect finished).

This commit reverts this behavior so looping sounds with
overlap-behavior "ignore" or "reset" can have their fade out
effect "smoothly" canceled by a PlaySound call.

To avoid that sounds owned by certain application get resumed back
even if that application was closed, we check if the player's
refcount is 0, and if so, it cannot be resumed, because no
application (bus name) is referring it.

https://phabricator.endlessm.com/T25659