server: Set countdown to None when the idle method gets called
Before, if ensure_release_countdown was called, and after ten
seconds, the release method was called, then self._countdown_id
kept with a value. So if you called cancel_countdown,
GLib.Source.remove attempted to remove an unexisting id.

Now, because we ensure that self._countdown_id is None after
the idle release function is called, then the call to
GLib.Source.remove will be never reached in this case.

https://phabricator.endlessm.com/T25991